Fix Windows compilation (cast pointer diff to long)

+++ pythran/pythonic/include/types/numpy_expr.hpp
@@ -98,7 +98,7 @@ namespace types
                      utils::index_sequence<I...>) const
     {
       std::initializer_list<long> distances{
-          (std::get<I>(iters_) - std::get<I>(other.iters_))...};
+          (static_cast<long>(std::get<I>(iters_) - std::get<I>(other.iters_)))...};
       return *std::max_element(distances.begin(), distances.end());
     }
